Add toggle to hide mount entry
